sentiment_analysis/requirements.txt
nltk==3.8.1
pandas==1.5.2
scikit-learn==1.2
numpy==1.24.0
matplotlib==3.6.3
requests==2.28.2
beautifulsoup4==4.11.1
tweepy==4.12.1
gensim==4.3.0
langdetect==1.0.9
python-dotenv==0.21.0